+++ tests/TreeFile.tests.ps1
@@ -8,14 +8,14 @@ Import-Module ([System.IO.Path]::Combine($PSScriptRoot, 'shared.psm1'))
 
 Describe 'TreeFile' {
     It 'Has a reference of the parent DirectoryInfo instance' {
-        $instance = $testPath | Get-PSTree | Where-Object { $_ -is [PSTree.TreeFile] } |
+        $instance = $testPath | Get-PSTree | Where-Object { $_ -is [PSTree.Nodes.TreeFile] } |
             Select-Object -First 1
 
         $instance.Directory | Should -BeOfType ([System.IO.DirectoryInfo])
     }
 
     It 'Has a reference of the parent directory path' {
-        $instance = $testPath | Get-PSTree | Where-Object { $_ -is [PSTree.TreeFile] } |
+        $instance = $testPath | Get-PSTree | Where-Object { $_ -is [PSTree.Nodes.TreeFile] } |
             Select-Object -First 1
 
         $instance.DirectoryName | Should -BeExactly (Get-Item $instance.FullName).DirectoryName